server/requirements.txt
Flask==1.0.3
Pillow==6.0.0
numpy==1.16.3
requests==2.21.0